+++ crawlserv/src/Main/Server.hpp
@@ -186,9 +186,9 @@ namespace crawlservpp::Main {
 
 		// getters
 		const std::string& getStatus() const;
-		unsigned long getUpTime() const;
-		unsigned long getActiveThreads() const;
-		unsigned long getActiveWorkers() const;
+		size_t getUpTime() const;
+		size_t getActiveThreads() const;
+		size_t getActiveWorkers() const;
 
 		// command function
 		bool tick();
@@ -318,11 +318,11 @@ namespace crawlservpp::Main {
 
 		ServerCommandResponse cmdDownload();
 
-		void cmdImport(ConnectionPtr connection, unsigned long threadIndex, const std::string& message);
-		void cmdMerge(ConnectionPtr connection, unsigned long threadIndex, const std::string& message);
-		void cmdExport(ConnectionPtr connection, unsigned long threadIndex, const std::string& message);
+		void cmdImport(ConnectionPtr connection, size_t threadIndex, const std::string& message);
+		void cmdMerge(ConnectionPtr connection, size_t threadIndex, const std::string& message);
+		void cmdExport(ConnectionPtr connection, size_t threadIndex, const std::string& message);
 
-		void cmdTestQuery(ConnectionPtr connection, unsigned long threadIndex, const std::string& message);
+		void cmdTestQuery(ConnectionPtr connection, size_t threadIndex, const std::string& message);
 
 		// private helper functions
 		static bool workerBegin(
@@ -331,7 +331,7 @@ namespace crawlservpp::Main {
 				ServerCommandResponse& response
 		);
 		void workerEnd(
-				unsigned long threadIndex,
+				size_t threadIndex,
 				ConnectionPtr connection,
 				const std::string& message,
 				const ServerCommandResponse& response
